requirements.txt
--extra-index-url https://artifactory.f5net.com/artifactory/api/pypi/f5-pypi-internal/simple/
pytest==3.2.2
mock==2.0.0
pexpect==4.7.0
PyYAML==5.1.2
pytest-cov==2.5.1
Jinja2==2.11.1
paramiko==2.6.0
requests~=2.24.0
isoparser==0.3
bigsuds==1.0.6
six==1.12.0
peewee==2.10.2
kombu==4.6.7
fluent-logger==0.9.4
deepdiff==5.0.2
pyparsing==2.4.7
click==7.1.2
fabric==2.5.0
gitpython==3.1.7
python-gnupg==0.4.6
f5-sdk==3.0.21
f5-sdk-python==0.9.2
ucs_modifier==0.3.3